Documentation & split
