+++ pubsub/Cargo.toml
@@ -10,6 +10,7 @@ keywords = ["jsonrpc", "json-rpc", "json", "rpc", "macros"]
 documentation = "https://ethcore.github.io/jsonrpc/jsonrpc_pubsub/index.html"
 
 [dependencies]
+log = "0.3"
 parking_lot = "0.3"
 jsonrpc-core = { version = "6.0", path = "../core" }
